SDK3.0 Release (#183)
* Makes Optimum compatible with the Poplar 3.0 SDK release.
* Enable dispatch tracing by default
* Replace the custom gather op in deberta with native unfold implementation
* Updated all the models for dispatch mode
* Enabled dictionary output on all models
* Replaced convnext layernorm with FP16 safe version.

+++ examples/speech-recognition/run_inference_ctc.py
@@ -108,7 +108,9 @@ def main():
     opts = ipu_config.to_options(for_inference=True)
     inference_model = poptorch.inferenceModel(ipu_model.half().eval(), options=opts)
 
-    sample_batch = {"input_values": torch.zeros([num_device_iterations, inference_args.max_samples])}
+    sample_batch = {
+        "input_values": torch.zeros([num_device_iterations, inference_args.max_samples], dtype=torch.float16)
+    }
 
     inference_model.compile(**sample_batch)
 
@@ -126,7 +128,7 @@ def main():
             length = input_values.size(1)
             x[i, :length] = input_values[0]
 
-        batch = {"input_values": x}
+        batch = {"input_values": x.to(dtype=torch.float16)}
 
         # take argmax and decode
         output = inference_model(**batch)
